refactor of imports, small changes, and css cleanup
