+++ package.json
@@ -39,11 +39,11 @@
 	},
 	"devDependencies": {
 		"@nabla/vite-plugin-eslint": "^1.4.1",
-		"@types/node": "^18.7.14",
-		"@typescript-eslint/eslint-plugin": "^5.36.1",
-		"@typescript-eslint/parser": "^5.36.1",
-		"@vitejs/plugin-vue": "^3.0.3",
-		"@vue/eslint-config-typescript": "^11.0.0",
+		"@types/node": "^18.7.15",
+		"@typescript-eslint/eslint-plugin": "^5.36.2",
+		"@typescript-eslint/parser": "^5.36.2",
+		"@vitejs/plugin-vue": "^3.1.0",
+		"@vue/eslint-config-typescript": "^11.0.1",
 		"autoprefixer": "^10.4.8",
 		"eslint": "^8.23.0",
 		"eslint-config-prettier": "^8.5.0",
@@ -55,8 +55,8 @@
 		"prettier-plugin-tailwindcss": "^0.1.13",
 		"tailwindcss": "^3.1.8",
 		"typescript": "^4.8.2",
-		"vite": "^3.0.9",
+		"vite": "^3.1.0",
 		"vue-eslint-parser": "^9.0.3",
-		"vue-tsc": "^0.40.5"
+		"vue-tsc": "^0.40.9"
 	}
 }
