+++ src/components/Sidebar.vue
@@ -2,7 +2,7 @@
 import Search from './SidebarComponents/Search.vue'
 import Options from './SidebarComponents/Options.vue'
 import Selection from './SidebarComponents/Selection.vue'
-import SiteInfo from './SiteInfo.vue'
+import SiteInfo from './SidebarComponents/SiteInfo.vue'
 import SaveImage from './SidebarComponents/SaveImage.vue'
 import { ref } from 'vue'
 import Export from './SidebarComponents/Export.vue'
@@ -20,12 +20,21 @@ function onCanRender() {
 </script>
 
 <template>
-	<div class="flex-col tw-sidebar-width fixed bg-[#303030] h-full">
+	<div
+		class="flex flex-col gap-3 tw-sidebar-width fixed bg-[#303030] h-full mt-2"
+	>
 		<SiteInfo />
-		<Selection @can-render-chart="onCanRender" />
+
+		<div class="mt-[-4px]">
+			<Selection @can-render-chart="onCanRender" />
+		</div>
+
 		<Export />
+
 		<SaveImage />
+
 		<Options v-if="showOptions" />
+
 		<Search />
 	</div>
 </template>
