+++ vite.config.ts
@@ -25,7 +25,8 @@ export default defineConfig({
 			'#src': path.resolve(__dirname, './src'),
 			'#assets': path.resolve(__dirname, './src/assets'),
 			'#components': path.resolve(__dirname, './src/components'),
-			'#types': path.resolve(__dirname, './src/types')
+			'#types': path.resolve(__dirname, './src/types'),
+			'#core': path.resolve(__dirname, './src/components/CoreComponents')
 		}
 	}
 })
